requirements.txt
pywebview>=5.3
playwright>=1.50.0
peewee>=3.17.0
python-docx>=1.1.0
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.0.0
